+++ Av1an/__init__.py
@@ -1,4 +1,3 @@
 #!/usr/bin/env python3
 
 from .arg_parse import arg_parsing
-from .encode import main_queue