+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - Default extra_splits set to 240 frames.
 - Extraction and Concatenation to copy all streams.
 - Scenes file to save total frame count, faster restart/start up with scenes file.
+- Fix concat on windows with ffmpeg